Add files via upload
Unit testing is the process of testing program components, such as methods or object classes. Unit includes Individual functions or methods within an object.
When testing object classes, test design is crucial to provide coverage of all of the features of the object. It should test all operations associated with the object; set and check the value of all attributes associated with the object; and put the object into all possible states.